Make registrations specify MobX-transaction explicitly
This is a new requirement from injectable-extension-for-mobx, as otherwise registrations could surprisingly cause untimely, and too numerous observations by computedInjectMany. The bad thing is the duplication, but I do currently consider this as favorable, as now a dev will always know that registrations of injectables with tokens may cause triggering of observations.

Co-authored-by: Janne Savolainen <janne.savolainen@live.fi>

Signed-off-by: Iku-turso <mikko.aspiala@gmail.com>